==> code/04_RichnessDepth.Rmd <==
---
title: "04_RichnessDepth"
output: html_document
date: "2025-02-12"
---

```{r}
library(rstanarm)
```

Plot # of ASVs and/or # of species against sampling distance from bottom

```{r}

load("/Users/stephaniematthews/Documents/Rockfish/Conservation_Letters/Data/Rockfish_phyloseq.RData")
field_curated_MDL <- subset_taxa(field_curated_MDL, Species != "Sebastes (unknown rockfish)")
sample_data(field_curated_MDL)$Basin[sample_data(field_curated_MDL)$Site %in% c("HC-8", "HC-9")] <- "AI"#move HC 8 and 9 to AI, since they're more similar to AI 
sample_data(field_curated_MDL)$Basin[sample_data(field_curated_MDL)$Basin %in% c("AI")] <- "Puget Sound"
sample_data(field_curated_MDL)$Basin[sample_data(field_curated_MDL)$Basin %in% c("HC")] <- "Hood Canal"

#nsamples(field_curated_MDL)
table(sample_data(field_curated_MDL)$Instrument)

sample_data(field_curated_MDL)$Observed_ASVs <- estimate_richness(field_curated_MDL, split= T, "Observed")$Observed
sample_data(field_curated_MDL)$Observed_Species <- estimate_richness(tax_glom(field_curated_MDL, "Species"), split= T, "Observed")$Observed
metadata_field_curated_MDL <- data.frame(sample_data(subset_samples(field_curated_MDL, Basin %in% c("Puget Sound", "Hood Canal"))))
metadata_field_curated_MDL <- metadata_field_curated_MDL %>% select(where(~!all(is.na(.x))))

metadata_field_curated_MDL$WaterColumnLoc <- "Subsurface"
metadata_field_curated_MDL$WaterColumnLoc[metadata_field_curated_MDL$Depth_m == 0] <- "Surface"
metadata_field_curated_MDL$WaterColumnLoc <- factor(metadata_field_curated_MDL$WaterColumnLoc, levels = c("Surface", "Subsurface"))

p <- ggplot(metadata_field_curated_MDL) + 
  geom_jitter(aes(x=Distance_from_bottom, y = Observed_ASVs, color = Instrument), width = 2, height = 0.2, alpha = 0.5, size = 3) + 
#  facet_wrap(.~PCR2.all.samples, scales = "free_x") + 
  xlab("Distance from bottom (m)") + 
  ylab("Sebastes ASVs")+
  ggtitle("") + 
    scale_color_manual(values = c("#104E8B", "#9BCD9B", "#EEAEEE"), name = "Platform") + 
  scale_y_continuous(breaks = c(0, 2, 4, 6, 8))
p

jpeg("~/Documents/Rockfish/Conservation_Letters/Figures/Fig4ASVs_vs_distancefrombottom.jpeg", width = 4, height = 3, units = "in", res = 300)
p
dev.off()

pdf(file = "~/Documents/Rockfish/Conservation_Letters/Figures/Fig4ASVs_vs_distancefrombottom.pdf", height =3, width = 4)
p 
dev.off()

p <-metadata_field_curated_MDL %>% 
  arrange(Instrument) %>%
 ggplot() + 
  geom_jitter(aes(x=Distance_from_bottom, y = Observed_Species, color = Instrument, shape = WaterColumnLoc), width = 1, height = 0.15, alpha = 0.45, size = 2) + 
#  facet_wrap(.~PCR2.all.samples, scales = "free_x") + 
  xlab("Distance from seafloor (m)") + 
  ylab("Sebastes species")+
  ggtitle("") + 
    scale_color_manual(values = c("#104E8B", "#9BCD9B", "#EEAEEE"), name = "Platform") + 
  scale_y_continuous(breaks = c(0, 1, 2, 3, 4, 5, 6, 8)) + 
  scale_shape_manual(values = c(17, 19), name = "Sampling\nlocation")
p

jpeg("~/Documents/Rockfish/Conservation_Letters/Figures/Fig4Species_vs_distancefrombottom.jpeg", width = 4, height = 3, units = "in", res = 300)
p
dev.off()

pdf(file = "~/Documents/Rockfish/Conservation_Letters/Figures/Fig4Species_vs_distancefrombottom.pdf", height = 3, width = 4)
p 
dev.off()

```

#number of species detected by sampling platform
```{r}
instrument_pa <- phyloseq_standardize_otu_abundance(merge_samples(field_single, group = "Instrument"), "pa")
instrument_pa_dataframe <- instrument_pa  %>%
  psmelt()  
instrument_pa_dataframe$Abundance[is.nan(instrument_pa_dataframe$Abundance)] <- 0
instrument_pa_dataframe$Abundance_binary[(instrument_pa_dataframe$Abundance) == 0] <- "Absent"
instrument_pa_dataframe$Abundance_binary[(instrument_pa_dataframe$Abundance) == 1] <- "Present"

#by site
instrument_pa_dataframe %>% 
  group_by(Instrument) %>%
  count(Abundance_binary) %>%
  print(n = 25)

```

## run poisson
```{r}

metadata_field_curated_MDL %>% 
  arrange(Instrument) %>%
 ggplot() + 
  geom_jitter(aes(x=Distance_from_bottom, y = Observed_Species, color = Instrument, shape = WaterColumnLoc), width = 1, height = 0.15, alpha = 0.45, size = 2) + 
  stat_smooth(aes(x=Distance_from_bottom, y = Observed_Species), method = "glm")



summary(m1 <- glm(Observed_Species ~ Distance_from_bottom, family="poisson", data=metadata_field_curated_MDL))

##cover to run stan model and calculate credibility interval
m1_bglm <- stan_glm(Observed_Species ~ Distance_from_bottom, family="poisson", data=metadata_field_curated_MDL)
summary(m1_bglm, digits = 2, prob=c(.025, .5, .975))

m1_bglm <- stan_glm(Observed_Species ~ Distance_from_bottom, 
                  data = metadata_field_curated_MDL,
                  family = poisson(link = "log"), #this is the default for poisson
                  seed = 123)


# Get coefficient estimates
summary(m1_bglm)

# Plot model diagnostics
plot(m1_bglm)

# Extract posterior samples
posterior_samples <- as.matrix(m1_bglm)

# get 95% CI for intercept and slope:
posterior_interval(m1_bglm)

#OR:
# intercept
quantile(posterior_samples[,1], probs = c(0.025, 0.975))
# slope
quantile(posterior_samples[,2], probs = c(0.025, 0.975))



metadata_field_curated_MDL$predicted <- posterior_epred(m1_bglm) %>% colMeans() 
metadata_field_curated_MDL$predicted_hi <- apply(posterior_epred(m1_bglm), MARGIN = 2, FUN = quantile, probs = 0.975)
metadata_field_curated_MDL$predicted_lo <- apply(posterior_epred(m1_bglm), MARGIN = 2, FUN = quantile, probs = 0.025)


## create the plot
ggplot(metadata_field_curated_MDL, aes(x = Distance_from_bottom, y = predicted)) +
  geom_point(aes(y = Observed_Species), alpha=.5, position=position_jitter(h=.2)) +
  geom_line(size = 1) +
  labs(x = "Distance from seafloor (m)", y = "Sebastes species")

p <- metadata_field_curated_MDL %>% 
 # arrange(Instrument) %>%
 ggplot() + 
  geom_jitter(aes(x=Distance_from_bottom, y = Observed_Species, color = Instrument, shape = WaterColumnLoc), width = 1, height = 0.15, alpha = 0.45, size = 2) + 
  geom_line(aes(x = Distance_from_bottom, y = predicted), size = 1, color = "grey40") +
  geom_ribbon(aes(x = Distance_from_bottom, ymin=predicted_lo, ymax=predicted_hi), linetype=2, alpha=0.3) +
  xlab("Distance from seafloor (m)") + 
  ylab("Sebastes species")+
   ggtitle("") + 
    scale_color_manual(values = c("#104E8B", "#9BCD9B", "#EEAEEE"), name = "Platform") + 
  scale_y_continuous(breaks = c(0, 1, 2, 3, 4, 5, 6, 8)) + 
  scale_shape_manual(values = c(17, 19), name = "Sampling\nlocation")
p 


jpeg("~/Documents/Rockfish/Conservation_Letters/Figures/Fig4Species_vs_distancefrombottom_lm.jpeg", width = 4, height = 3, units = "in", res = 300)
p
dev.off()

pdf(file = "~/Documents/Rockfish/Conservation_Letters/Figures/Fig4Species_vs_distancefrombottom_lm.pdf", height = 3, width = 4)
p 
dev.off()


```
